Warning patrol
git-svn-id: http://svn.boost.org/svn/boost/trunk@70017 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ libs/random/test/test_bernoulli.cpp
@@ -25,7 +25,7 @@
 bool do_test(double p, long long max) {
     std::cout << "running bernoulli(" << p << ")" << " " << max << " times: " << std::flush;
 
-    boost::math::binomial expected(max, p);
+    boost::math::binomial expected(static_cast<double>(max), p);
     
     boost::random::bernoulli_distribution<> dist(p);
     boost::mt19937 gen;
