+++ libs/random/test/test_geometric.cpp
@@ -12,6 +12,7 @@
 #include <boost/random/geometric_distribution.hpp>
 #include <boost/random/uniform_real.hpp>
 #include <boost/math/distributions/geometric.hpp>
+#include <boost/numeric/conversion/cast.hpp>
 
 #define BOOST_RANDOM_DISTRIBUTION boost::random::geometric_distribution<>
 #define BOOST_RANDOM_DISTRIBUTION_NAME geometric
@@ -20,6 +21,6 @@
 #define BOOST_RANDOM_ARG1_NAME p
 #define BOOST_RANDOM_ARG1_DEFAULT 0.5
 #define BOOST_RANDOM_ARG1_DISTRIBUTION(n) boost::uniform_real<>(0.0001, 0.9999)
-#define BOOST_RANDOM_DISTRIBUTION_MAX (-5 / std::log(1-p))
+#define BOOST_RANDOM_DISTRIBUTION_MAX boost::numeric_cast<int>(-5 / std::log(1-p))
 
 #include "test_real_distribution.ipp"
